refactor: IntensityIcon 系モデルを専用ファイルへ分離
